Add a binary for running background jobs and monitoring their status
These binaries are pretty simple (and a bit spaghetti) for the time
being, since we have so little actually using the background job
framework. The runner is meant to be run on a worker dyno continuously,
while the monitor should be run by a cron-like tool roughly every 5
minutes.

We should move `update-downloads` to be scheduled as well, since it
spends so little time actually working.