`cargo update -p log:0.3.8`
I ran into this issue while trying to debug the login issue.  We have a
few dependencies that pull in an older version of `log`, which means
that, for instance, `RUST_LOG=oauth2=debug` doesn't work.  `log:0.3.9`
provides a semver compatibile wrapper around `log:0.4` which allows
logging to work for all our dependencies which support it.