Yanking only updates the database after the index is updated
Previously, we'd immediately update the database and then queue the
index update to be run later. Jobs aren't guaranteed to run in the
order they were queued (though in practice they likely will). This means
that if someone yanked and unyanked a crate in rapid succession, the
database may end up permanently out of sync with the index.

With this change, the final result may still be out of order, but the
database will be in sync with the index whatever the outcome is. Since
rapidly yanking and unyanking isn't something we expect users to do, and
even if they did jobs should be running in order under normal
circumstances, I don't think we need to do anything to ensure more
consistency than this.