Add a `include_yanked` parameter to /api/v1/crates
Add a URL parameter to the `/api/v1/crates` endpoint that specifies
whether or not the endpoint should return crates that have been fully
yanked - i.e., where every version has been yanked.

If not specified, this defaults to `true` because we have typically
always returned all matching crates, whether or not their versions have
been yanked.

This makes it possible to display only non-yanked crates at various
places in the UI, and is especially intended for the user and team pages
to display only non-yanked crates.