Allow profiling jemalloc
This builds jemalloc with profiling enabled, and preserves debug symbols
in the release binary (note: `debug = true` does not disable
optimizations). This commit by itself doesn't actually perform any
profiling, but it allows it to be controlled by an environment variable.

At this point I cannot get any useful information about the memory issue
we're seeing in production locally. We also can't reproduce it in a
staging environment. There's *something* about our production traffic
patterns that we're missing, so the next step is to start looking at
dumps of production memory usage.

My intention is to start enabling sampling based profiling in
production, and copy some of the dumps off the server to examine
locally.