Fix pool size, don't crash if we can't get a connection
When testing in staging, the runner started crashing on boot. This was
because we only set the connection pool size to 1 when we meant to set
it to 2 (one for the runner, one for the worker thread). So in each
loop, if all jobs took greater than 1 second to run, the runner would
crash.

This fixes the pool size, and also does not return an error if no
database connection could be retrieved from the pool.