Check if a crate is already yanked/unyanked in the job
Right now we check if a crate is already yanked/unyanked before enqueing
the job at all. This means that if you try to undo a yank before the
yank finishes running, we won't enqueue the unyank at all. This isn't an
expected scenario, and we still might not do the unyank if the jobs are
run out of order, but this means that we should always end up in the
expected state under normal operation.