Move the background jobs to another crate
I wrote this code intending to extract it, and I'm planning to release
it after a few more changes which aren't relevant to crates.io. The code
right now is largely the same as what was here before, the main
difference being the use of a different error type.

The error handling will probably churn some more in the near future as I
figure out what I want the error handling story to look like in that
crate