Display errors during login
The code was serializing an entire response object, which didn't
actually contain the errors that came back from the API, so we'd never
show anything other than "Failed to log in". With this change we now
show the error returned by the API if it returned any. Notably, if a
user tries to create a new account while we're in read only mode, they
will now see an error saying that.