Remove `version_downloads.id`
This column is no longer being used, and can be safely removed. Rolling
back this migration in production requires creating indexes
concurrently, which Diesel CLI doesn't support, so rolling back this
migration in production will require queries to be run manually. A
blocking version is included for development and testing