Move README renderings off the web server
Rendering the README is a process that can take an arbitrarily long
amount of time, and ends with uploading a file to S3 -- which also can
take an arbitrarily long period of time. Since there are no README
contents which we consider invalid (which I've indicated by changing the
return type of the rendering functions to `String` instead of `Result`),
there's no reason for us to do any of this work on the web server.

Just like the index updates, we can offload this to our worker to be
completed at some point in the future. This does mean that a crate will
render in our UI before the README has finished rendering. During this
brief period, we will behave the same as if the crate had no README
file, showing the description instead.

Eventually I think it'd make sense to have this job behave like the
`render-readmes` worker, where it pulls the tar file down and reads the
actual readme file. This would mean that Cargo can stop sending us the
README as its own field when its already sending us a tar file which
contains the README. I've left it alone for now though, since there's no
benefit while Cargo is already sending us the README, and we can't have
it stop doing that until we move to publish v2.

A few implementation notes:

- The `readme` column on `crates` was never used for anything in Rust,
  but is used on the DB side to generate the search index. I've left the
  column in place, but removed most uses of it on the Rust side.
- The `readme_file` column on `crates` never used for anything. This
  commit does not include the migration, which must be deployed
  separately. This commit only includes code which works with or without
  the migration for deployment.
- `background_jobs::Environment` gained two fields that are just
  straight up copied from `App`. I've opted to do this rather than just
  giving it `App`, since I do want to separate "the stuff you need to do
  background jobs" from "the stuff you need to be the web server". Since
  we do eventually plan on not having s3 uploads happen on the server, I
  want to bundle all of this into a single field and just have it live
  on `background_jobs::Environment` eventually.
- I did some general cleanup of `krate::publish` while I was in there,
  removed a lot of pre-1.0 style code.
- I've left the `render-readmes` binary alone for now since it's
  untested. I'll change it to just queue up jobs once we get to the
  point where jobs are doing all the tar file work as well.