Remove usages of the now-unused `license` col on the `crates` table
This is a big change!  `cargo check` says this is fine, but some
manual verification and testing is probably a good idea.  (`cargo test`
does pass locally for me, but that only tells half the story.)

As part of (GitHub) #736, this commit not only removes the commitment
of license information into the database, but it also removes the
handling in the controller.  For silent backwards compatibility, it
might be better to simply drop the license field, for example if it's
being uploaded by outdated clients.  I am not sure, however, and will
probably need feedback on this.

(A commit running `cargo fmt` was also rebased into this one.)

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>