backend: Add user email preference storing
This commit is the first step in the process to send crate owners an
email notification when a new version of one of their crates is
published. A database migration adds a `email_notifications` column to
the `crate_owners` table, and thus the property was added to the
corresponding `CrateOwner` struct. This new property is defaulted to
`true`.

Because a user may not want to receive a version publish notification
for all of their crates, an API endpoint was added to allow them to
toggle email notifications for each crate. The front end implementation
will be in a forthcoming commit, as well as the actual sending of these
notifications.