Link directly to cargo install page
This fixes a papercut where after clicking on the button, clicking the
back button causes the user to immediately redirect again.

Additionally, on Firefox 71, it was not possible to long-click on the
back button to return to crates.io (as the only entry in the list was
the redirect.  On Chromium 79 I had no such issue.