make title dynamic again
There is a bunch of code that sets the page title dynamically,
but it is all being overridden by a title tag in the main
app/index.html template (the DOM is only allowed to have
a single title tag in it, so subsequent title tags are
ignored by all browsers).

There is probably a better way to do this (using fastboot
to fill in the correct title on the backend would be ideal)
but this patch should be enough to restore the dynamic
titles for browsers with javascript enabled.