Set the TrustedUserId in the extensions when logging in
Fixes #2166.

In #2077 (d0b9bccc4cf80d3e311696a7a9e62ee469cce892), we changed the type
of the authentication data stored in the extensions from `User` to
`TrustedUserId`, but we missed this spot :( We're allowed to insert
anything we want into the extensions, but if we use a different type
than what we use when looking up data in the extensions, it won't find
what we inserted.