index: Move data loading back into the controller
This partially reverts commit a34f6c47a70a00f26bf1c1e20718aa1076f3a967.

The goal of this commit is to display loading spinners for the sections that are still loading, but display the rest of the content of the frontpage already, while the summary data is still loading.