tests/krate/publish: Fix `new_krate_tarball_with_hard_links()`
The test was receiving "too large when decompressed" because of a broken header checksum, which was mistakenly converted into this error message, because we assume that every `entries()` error is coming from our `LimitErrorReader`.  This commit moves the checksum calculation to the right place, which causes the test to now test the correct condition in our `verify_tarball()` function.