Boot nginx from the backend app to fix graceful shutdown in production
In production we aren't gracefully shutting down as expected. nginx
does a quick shutdown when receiving a `SIGTERM` and once it exits, it
appears that our process is aborted.

Instead of spawning the backend from the shell script, the backend
process is now responsible for spawning the shell script. This allows
the backend to finish its shutdown work even if the script has already
exited.

`cat` is passed to the nginx buildpack script as our "server", which
will block on `STDIN` so that `nginx` will run until receiving
`SIGTERM`.