add support for running tests with the real database pool
Some tests need access to a "real" production database pool, but that
means we can't rely on begin_test_transaction(), as that only works
inside a single connection.

This commit adds support creating a brand new schema for each test, and
adds a way for tests to opt into that. This is not enabled for all
tests, as creating a new schema and running migrations on it takes
around 1.5 seconds on my computer.