Disable page offset blocking if UA blocklist is unset
This fixes a bug where if `WEB_PAGE_OFFSET_UA_BLOCKLIST` was not set,
then all user-agents were blocked for large page offsets. This was
because all string contain the empty string.

Now, no blocking occurs if the variable is unset or empty. If the comma
separated list contains an empty string, then all user-agents will be
blocked.