mirage: Untangle `crate.id` and `crate.name`
The `id` is rarely used and exposed on the API, but there are a few cases where it is indeed used. The `id` is meant to be a `i32` though, so having it equal to the `name` of the crate was wrong. This commit fixes the issue by having the `id` as a regular numeric ID and using `name` everywhere else.